+++ src/V1/Tests/TestFiles/UserConfirmEmailRuleTest.cs
@@ -5,7 +5,7 @@ using ServiceQuery;
 namespace ServiceBricks.Xunit.Integration
 {
     [Collection(ServiceBricks.Xunit.Constants.SERVICEBRICKS_COLLECTION_NAME)]
-    public abstract class UserConfirmEmailRuleTestBase : IAsyncDisposable
+    public abstract class UserConfirmEmailRuleTestBase : IDisposable, IAsyncDisposable
     {
         public UserConfirmEmailRuleTestBase()
         {
@@ -13,12 +13,40 @@ namespace ServiceBricks.Xunit.Integration
 
         public ISystemManager SystemManager { get; set; }
 
-        public async ValueTask DisposeAsync()
+        public void Dispose()
         {
-            await CleanupDependencies();
+            CleanupDependencies();
         }
 
-        protected virtual async Task CleanupDependencies()
+        public async ValueTask DisposeAsync()
+        {
+            await CleanupDependenciesAsync();
+        }
+
+        protected virtual void CleanupDependencies()
+        {
+            string userstoragekey = UserStorageKey;
+
+            // Cleanup Audits
+            var audituserservice = SystemManager.ServiceProvider.GetRequiredService<IUserAuditApiService>();
+            var auditquery = new ServiceQueryRequestBuilder().IsEqual(nameof(UserAuditDto.UserStorageKey), userstoragekey).Build();
+            var respAudits = audituserservice.Query(auditquery);
+            foreach (var item in respAudits.Item.List)
+                audituserservice.Delete(item.StorageKey);
+
+            // Cleanup UserRoles
+            var userroleservice = SystemManager.ServiceProvider.GetRequiredService<IApiService<UserRoleDto>>();
+            var roleq = new ServiceQueryRequestBuilder().IsEqual(nameof(UserRoleDto.UserStorageKey), userstoragekey).Build();
+            var respUserRoles = userroleservice.Query(roleq);
+            foreach (var item in respUserRoles.Item.List)
+                userroleservice.Delete(item.StorageKey);
+
+            // Cleanup User
+            var userservice = SystemManager.ServiceProvider.GetRequiredService<IApiService<UserDto>>();
+            userservice.Delete(userstoragekey);
+        }
+
+        protected virtual async Task CleanupDependenciesAsync()
         {
             string userstoragekey = UserStorageKey;